Revert "Bernardo"
