+++ Website/public/css/header.css
@@ -10,10 +10,10 @@
   display: flex;
   padding: 100px 50px;
   text-align: center;
-  width: 40%;
+  width: 100%;
   height: 100vh;
-  background: #a69c94;
-  transition: all 0.6s ease-out;
+  background: #A69C94;
+  transition: 0.5s;
   z-index: 999;
   border-top: 1px solid rgba(0,0,0,0.2);
   flex-direction: column;
@@ -27,7 +27,6 @@
 
 .navbar-content.active{
   left: 0;
-  background: rgb(64, 56, 49);
 }
 
 .hamburgesa, .hamburgesa_close {
