+++ Website/public/css/list.css
@@ -25,7 +25,6 @@
     margin: 3px 3px 0 3px;
     object-fit: cover;
     border-radius: 3%;
-    box-shadow:  0 0 2px 0 rgba(0,0,0, 0.6);
 }
 
 .title-oferta{
@@ -52,7 +51,7 @@
     font-size: 1rem;
     list-style: none;
     font-family: 'Trebuchet MS';
-    padding: 10px 0 3px 0;
+    padding: 3px 0;
     display: flex;
     flex-wrap: wrap;
     align-items: center;
@@ -85,7 +84,7 @@
 }
 
 .main_section ul li:nth-child(2){
-   margin: 7px 3px;
+   margin: 5px 0;
 }
 
 @media screen and (min-width: 768px) {
