+++ Website/public/js/burgerMenu.js
@@ -1,7 +1,8 @@
 const toggle = document.querySelector('.hamburgesa');
+const toggle_close= document.querySelector('.hamburgesa_close');
 const menuToggle = document.querySelector('.navbar-content');
 
     toggle.addEventListener('click', () => menuToggle.classList.add('active'))
-    menuToggle.addEventListener('mouseleave', () => menuToggle.classList.remove('active'))
+    toggle_close.addEventListener('click', () => menuToggle.classList.remove('active'))
