+++ Website/src/middlewares/user.js
@@ -20,6 +20,7 @@ const middleware = (req, res, next) => {
     
     })
     .catch(error => res.send(error))
-  
+    
+   
 }
 module.exports = middleware
