+++ Website/src/routes/users.js
@@ -4,7 +4,15 @@ const router = express.Router();
 const login = require('../middlewares/login');
 const save = require('../middlewares/save')
 const access = require('../middlewares/access')
-const upload = require('../middlewares/uploadUserImg');
+const multer = require('multer');
+const path = require('path');
+
+const storage = multer.diskStorage({
+    destination: (req, file, cb) => cb(null, path.resolve(__dirname, '../../public/img/Usuarios')),
+    filename: (req, file, cb) => cb(null, file.fieldname + '-' + Date.now() + path.extname(file.originalname))
+})
+
+const upload = multer({storage});
 
 router.get("/login", users.login);
 router.get("/register", users.register);
@@ -14,6 +22,6 @@ router.post("/save",[save], users.save);
 router.post("/access", [login], users.access);
 router.post("/logout",users.logout);
 
-router.put("/profile", upload.single('avatar'), users.uploadAvatar)
+router.put("/profile",[upload.single('avatar')], users.uploadAvatar)
 
 module.exports = router;
