+++ Website/src/views/partials/head.ejs
@@ -8,5 +8,4 @@
     <title> <%= title %> | Fashion Sale </title>
     <%- include('./styles') %> 
     <%- include('./libs') %> 
-    <script src="//cdn.jsdelivr.net/npm/sweetalert2@11"></script>
 </head>
