+++ Website/src/views/partials/header.ejs
@@ -14,6 +14,9 @@
   </span>
 
   <nav class="navbar-content">
+    <span class="hamburgesa_close">
+      <i class="fa-solid fa-bars hamburgesa"></i>
+    </span>
     <a href="/products/">Productos</a>
     <a href="/products/">Hombre</a>
     <a href="/products/">Mujer</a>
