+++ Website/src/views/products/productDetail.ejs
@@ -93,8 +93,8 @@
               <p>Talle modelo <%= product.category %>:
                 <ul class="listaTalles">
                  <% for( let i = 0; i < product.sizes.length; i++ ) { %>
-                  <li>Talle <strong><%= product.sizes[i].size %></strong> hay un total de <%= product.sizes[i].units %>
-                    <% if (product.sizes[i].units == 1) { %>
+                  <li>Talle <strong><%= product.sizes[i].sizes %></strong> hay un total de <%= unidades[i].units %>
+                    <% if (unidades[i].units == 1) { %>
                      unidad.
                     <% } else{%>
                       unidades.</li>
