+++ Website/src/views/users/login.ejs
@@ -1,3 +1,5 @@
+<!-- ACLARACIÓN: ESTO ES TEMPORAL, ES MERA Y EXCLUSIVAMENTE PARA DEMOSTRAR QUE SABEMOS COMO CREAR UN LOGIN, PERO EN EL FUTURO LO VAMOS A TENER EN EL MENÚ COLAPSABLE EN EL HEADER (VEASE BOCETO) -->
+
 <!DOCTYPE html>
 <html lang="es">
   <%- include('../partials/head') %> 
@@ -15,13 +17,14 @@
               <% if (locals.old && old.email) { %>
                value="<%= old.email %>"
               <% } else {%>
-               placeholder="Ingresa tu email" 
-              <% } %> />
+               placeholder="email" 
+              <% } %> 
+               required/>
             </fieldset>
             <fieldset>
               <legend class="no-show">Iniciar Sesión</legend>
               <label class="no-show" for="password">Contraseña</label>
-              <input type="password" name="password" id="pass" placeholder="Ingresa tu contraseña" />
+              <input type="password" name="password" id="pass" placeholder="contraseña" required/>
             </fieldset>
             <fieldset>
               <button type="submit" id="login-button">Iniciar Sesión</button>
@@ -35,7 +38,6 @@
               </label>
             </fieldset>
           </form>
-          <ul id="errorsList"></ul>
           <% if (locals.errors) { %>
             <% if (errors.email) { %>
                 <p><%= errors.email.msg %></p>
@@ -46,7 +48,7 @@
         <% } %>
         </article>
       </section>
-      <script src="/js/login.js"></script>
+      
     </main>
     <%- include('../partials/footer') %> 
   </body>
