+++ Website/src/views/users/register.ejs
@@ -20,7 +20,7 @@
               <% } else {%>
               placeholder="Nombre" 
               <% }%>
-              />
+              required/>
             </fieldset>
             <fieldset>
               <input class="controls" type="text" name="lastName" id="lastName"
@@ -29,7 +29,7 @@
               <% } else { %>
                 placeholder="Apellido" 
               <% } %>   
-              />
+              required />
             </fieldset>
             <fieldset>
               <input class="controls" type="email" name="email" id="email" placeholder="Email" />
@@ -40,8 +40,7 @@
             <fieldset>
               <button class="button" type="submit">Continuar</button>
             </fieldset>
-            <article>        
-              <ul id="errorsList"></ul>       
+            <article>               
               <% if (locals.errors) { %>
                 <% if (errors.email) { %>
                     <p><%= errors.email.msg %></p>
@@ -59,7 +58,6 @@
             </article>
           </form>
         </section>
-        
       </main>
       <%- include('../partials/footer') %>
   </body>
